chore: move all vega commands to a separte package
