+++ cmd/vega/commands/blockexplorer.go
@@ -10,26 +10,30 @@
 // of this software will be governed by version 3 or later of the GNU General
 // Public License.
 
-package main
+package commands
 
 import (
 	"context"
+	"os"
 
-	"code.vegaprotocol.io/vega/cmd/vega/verify"
-
+	cmd "code.vegaprotocol.io/vega/cmd/blockexplorer/commands"
 	"github.com/jessevdk/go-flags"
 )
 
-type VerifyCmd struct {
-	Asset   verify.AssetCmd   `command:"passet" description:"verify the payload of an asset proposal"`
-	Genesis verify.GenesisCmd `command:"genesis" description:"verify the appstate of a genesis file"`
+type blockExplorerCmd struct{}
+
+func (opts *blockExplorerCmd) Execute(_ []string) error {
+	os.Args = os.Args[1:]
+	return cmd.Execute(context.Background())
 }
 
-var verifyCmd VerifyCmd
+func BlockExplorer(ctx context.Context, parser *flags.Parser) error {
+	_, err := parser.AddCommand(
+		"blockexplorer",
+		"The vega block explorer backend",
+		"The vega block explorer backend",
+		&blockExplorerCmd{},
+	)
 
-func Verify(ctx context.Context, parser *flags.Parser) error {
-	verifyCmd = VerifyCmd{}
-
-	_, err := parser.AddCommand("verify", "Verify Vega payloads or genesis appstate", "", &verifyCmd)
 	return err
 }
