+++ cmd/vega/commands/faucet/init.go
