+++ cmd/vega/commands/nodewallet/generate.go
