+++ cmd/vega/commands/nodewallet/show.go
