+++ cmd/vega/commands/nodewallet/verify.go
