+++ cmd/vega/commands/paths/paths.go
