+++ cmd/vega/commands/query/accounts.go
