+++ cmd/vega/commands/query/client.go
