+++ cmd/vega/commands/verify.go
@@ -10,30 +10,26 @@
 // of this software will be governed by version 3 or later of the GNU General
 // Public License.
 
-package main
+package commands
 
 import (
 	"context"
-	"os"
 
-	cmd "code.vegaprotocol.io/vega/cmd/data-node/commands"
+	"code.vegaprotocol.io/vega/cmd/vega/commands/verify"
+
 	"github.com/jessevdk/go-flags"
 )
 
-type datanodeCmd struct{}
-
-func (opts *datanodeCmd) Execute(_ []string) error {
-	os.Args = os.Args[1:]
-	return cmd.Execute(context.Background())
+type VerifyCmd struct {
+	Asset   verify.AssetCmd   `command:"passet" description:"verify the payload of an asset proposal"`
+	Genesis verify.GenesisCmd `command:"genesis" description:"verify the appstate of a genesis file"`
 }
 
-func Datanode(ctx context.Context, parser *flags.Parser) error {
-	_, err := parser.AddCommand(
-		"datanode",
-		"The vega data node",
-		"The vega data node",
-		&datanodeCmd{},
-	)
+var verifyCmd VerifyCmd
 
+func Verify(ctx context.Context, parser *flags.Parser) error {
+	verifyCmd = VerifyCmd{}
+
+	_, err := parser.AddCommand("verify", "Verify Vega payloads or genesis appstate", "", &verifyCmd)
 	return err
 }
