+++ cmd/vega/commands/verify/genesis.go
